+++ packages/search-adapter/src/empathy/__tests__/request-mappers/empathy-search-request-mapper.spec.ts
@@ -13,11 +13,13 @@ const mockedRelatedTagsQueryMapper = {
     relatedTags.length > 0 ? relatedTags.reduce((chain, rt) => `${ chain } ${ rt.tag }`, query).trim() : query)
 };
 const mockedQueryMapper = { map: jest.fn((_rawQuery: string, query: string) => query) };
-const mockedFiltersMapper = { map: jest.fn(() => ['size=xxl']) };
+const mockedFiltersMapper = { map: jest.fn(() => []) };
+const mockedSortMapper = { map: jest.fn((sort: string) => sort) };
 
 container.bind(DEPENDENCIES.RequestMappers.Parameters.query).toConstantValue(mockedRelatedTagsQueryMapper);
 container.bind(DEPENDENCIES.RequestMappers.Parameters.query).toConstantValue(mockedQueryMapper);
 container.bind(DEPENDENCIES.RequestMappers.Parameters.filters).toConstantValue(mockedFiltersMapper);
+container.bind(DEPENDENCIES.RequestMappers.Parameters.sort).toConstantValue(mockedSortMapper);
 let mapper: EmpathySearchRequestMapper;
 const emptyContext: RequestMapperContext = { feature: '', url: '', requestOptions: {} };
 
@@ -58,4 +60,22 @@ it('Does not modify the query if no related tags are passed', () => {
   mapper.map(rawRequest, request, emptyContext);
 
   expect(request.q).toEqual(query);
+});
+
+it('Uses the provided sort mappers', () => {
+  const rawRequest: SearchRequest = {
+    sort: 'price asc',
+    query: 'milk',
+    origin: 'default'
+  };
+  const request = {} as EmpathySearchRequest;
+
+  expect(mapper.map(rawRequest, request, emptyContext)).toEqual({
+    filter: [],
+    sort: 'price asc',
+    q: 'milk',
+    origin: 'default'
+  });
+
+  expect(mockedSortMapper.map).toHaveBeenCalledWith('price asc', '', emptyContext);
 });