+++ packages/search-adapter/src/empathy/mappers/index.ts
@@ -1,6 +1,5 @@
 export * from './request';
 export * from './response';
 
-export * from './mappers.const';
 export * from './pipe-mappers';
 export * from './response.mappers';