+++ packages/search-adapter/src/empathy/mappers/request/index.ts
@@ -2,6 +2,7 @@ export * from './params/empathy-request-filters.mapper';
 export * from './params/empathy-request-filters-solr-syntax.mapper';
 export * from './params/empathy-request-query.mapper';
 export * from './params/empathy-request-related-tags-query.mapper';
+export * from './params/empathy-request-sort.mapper';
 
 export * from './empathy-clicks-recommendations-request.mapper';
 export * from './empathy-queries-recommendations-request.mapper';