+++ packages/search-adapter/src/empathy/mappers/response/filters/empathy-boolean-filter.mapper.ts
@@ -1,0 +1,21 @@
+import { BooleanFilter, Filter } from '@empathy/search-types';
+import { injectable } from 'inversify';
+import { ResponseMapper, ResponseMapperContext } from '../../../empathy-adapter.types';
+import { EmpathyFilter } from '../../../models';
+
+/**
+ * Generic Empathy BooleanFilter mapper.
+ *
+ * @public
+ */
+@injectable()
+export class EmpathyBooleanFilterMapper implements ResponseMapper<EmpathyFilter, Filter> {
+
+  map(rawFilter: EmpathyFilter, filter: BooleanFilter, context: ResponseMapperContext): BooleanFilter {
+    return Object.assign<BooleanFilter, Partial<BooleanFilter>>(filter, {
+      selected: rawFilter.selected ?? false,
+      value: rawFilter.filter,
+      totalResults: rawFilter.count
+    });
+  }
+}